requirements.txt
pymysql==1.0.2
sqlalchemy<1.4.0
flask_sqlalchemy==2.5.1
flask_migrate==2.7.0
sqlalchemy_utils==0.36.8
flask_script==2.0.6
flask_jwt_extended==4.4.2